:truck: move instrumentation classes to separate file
